requirements.txt
Pillow
pyjade
requests